# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 学习状态智能助手 API（板块一 + 板块二 MVP）
  version: 1.8.0
  description: |
    中学学情导学与状态评估 API。

    来源：`docs/api-design-unified.md`（唯一生效版本），对应 `PRD-学习状态智能助手.md`。
    本文件严格按该设计文档生成，未引入文档之外的字段。

    主流程：创建目标 ① → 生成计划 ② → 提交学习记录 ③ → 计算状态 ④ → 返回建议 ⑤（外加周期复盘与反馈回路）。

    约定要点：
    - 所有资源以当前 token 用户为作用域，不设 userId 路径参数，不存在跨用户读取接口。
    - 写接口可选携带 `Idempotency-Key` 头，24 小时内重复键直接返回首次结果。
    - 建议与复盘为异步生成：POST 返回 202 + `generation.status = pending`，前端轮询 GET 获取结果。
    - 监护人授权失效时所有写接口返回 403 `GUARDIAN_AUTHORIZATION_EXPIRED`，读接口不受影响。

    阶段口径：pilot。本文档为前后端与 QA 的唯一契约来源。项目当前处于 pilot 阶段（小范围试用验证，非 beta/正式发布），由学生团队维持。板块三契约已于 M1 转正（口径见 docs/module3-m0-decision-proposal.md v1.7 §4.11）。

    **v1.6.0（重构 M0 · 契约冻结）**——本次为重构第一阶段的契约增量，未改动任何既有字段名：
    - **稳定用户 ID 落地（D59）**：`User.userId` 语义明确为稳定内部主键（`u_` 前缀），与登录凭证
      解耦；`AuthMeResponse.user` 增补可选 `userId`（`email` 保留，前端展示仍用它）。
    - **新增 M0 实体 schema**：Exam / Collection / CollectionItem / UsageLedgerEntry / InviteCode /
      ViolationLog / ErrorReport / Medal / UserProfileEntry / TopicSummary / Explanation /
      ChatSession / ChatRawMessage。⚠️ 这些 schema **尚未挂到任何 path**，接口由各板块按里程碑补齐；
      字段名以本节为准，各板块不得另起。
    - **题本升格（D48）**：新增 `ErrorCause` / `ErrorIntent` 两个正交维度，`ErrorRecord` 系列扩展
      `errorCause` / `intent` / `sourceExamId` 三个可选字段（`errorType` 保留为自由文本，向后兼容）。
    - **目标（D6 / D49）**：`Goal` / `GoalSummary` / `GoalCreate` / `GoalUpdate` 增补
      `parentGoalId`（父子树）、`examId` + `targetScore`（引用考试）。`status` 的 enum 未改动。
...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /recommendations:
    post:
      tags: [个性化建议]
      operationId: createRecommendation
      summary: 手动请求生成建议
      description: |
        常规路径由 POST /learning-records 自动创建，本接口仅用于手动刷新或阶段回顾；超频返回 429
        （PRD 6.4 控成本）。异步生成：返回 202 + generation.status = pending。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationCreate'
            example:
              scene: post_session
              subject: SX
              recordId: r_88012
      responses:
        '202':
          description: 生成任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationPending'
              example:
                recommendationId: rec_20301
                scene: post_session
                subject: SX
                generation:
                  status: pending
                createdAt: '2026-08-16T19:46:02+08:00'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/GuardianAuthorizationExpired'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags: [个性化建议]
      operationId: listRecommendations
      summary: 建议列表
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: scene
          in: query
          required: false
          description: 场景过滤
          schema:
            $ref: '#/components/schemas/RecScene'
        - name: subject
          in: query
          required: false
          description: 学科过滤
          schema:
            $ref: '#/components/schemas/Subject'
        - name: status
          in: query
          required: false
          description: 默认 ready
          schema:
            type: string
            enum: [ready, all]
            default: ready
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 建议列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationList'
              example:
                items:
                  - recommendationId: rec_20301
                    scene: post_session
                    subject: SX
                    generation:
                      status: ready
                      source: llm
                      completedAt: '2026-08-16T19:46:09+08:00'
                    items:
                      - title: 把单次时长压到 25 分钟
                        content: 这次函数练了 45 分钟但中断了 3 次，后半程正确率明显掉下来了。下次试试练 25 分钟就停，中间歇 5 分钟。
                    basedOn:
                      assessmentId: a_7742
                      recordId: r_88012
                      stateLabel: fatigue_warning
                      explain: 依据最近 7 次数学记录的疲劳自评与正确率变化
                    feedback: null
                pagination:
                  page: 1
                  pageSize: 20
                  total: 12
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /recommendations/{recommendationId}:
    get:
      tags: [个性化建议]
      operationId: getRecommendation
      summary: 获取建议内容 ⑤
      description: |
        生成中时 generation.status 为 pending、items 为 null。
        LLM 失败走兜底时 generation.source 为 template，其余结构完全一致，前端无需第二套渲染逻辑
        （对应 PRD 9 验收标准「API 全部不可用时核心流程不受影响」）。
      parameters:
        - $ref: '#/components/parameters/RecommendationId'
      responses:
        '200':
          description: 建议详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recommendation'
              examples:
                ready_llm:
                  summary: 生成完成（LLM）
                  value:
                    recommendationId: rec_20301
                    scene: post_session
                    subject: SX
                    generation:
                      status: ready
                      source: llm
                      completedAt: '2026-08-16T19:46:09+08:00'
                    items:
                      - title: 把单次时长压到 25 分钟
                        content: 这次函数练了 45 分钟但中断了 3 次，后半程正确率明显掉下来了。下次试试练 25 分钟就停，中间歇 5 分钟。
                      - title: 先巩固再上新
                        content: 今天难度感受偏难，暂时别急着推进新内容，明天先把函数图像的已掌握题型过一遍找回手感。
                    basedOn:
                      assessmentId: a_7742
                      recordId: r_88012
                      stateLabel: fatigue_warning
                      explain: 依据最近 7 次数学记录的疲劳自评与正确率变化
                    feedback: null
                pending:
                  summary: 生成中
                  value:
                    recommendationId: rec_20301
                    scene: post_session
                    subject: SX
                    generation:
                      status: pending
                    items: null
                    feedback: null
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /recommendation-content:
    get:
      tags: [个性化建议]
      operationId: getRecommendationContent
      summary: 导学计划页学习内容推荐（LLM 驱动）
      responses:
        '200':
          description: 推荐内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationContent'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==================== 板块三：群体匿名参照（M1 契约转正 · 口径 module3-m0-decision-proposal.md v1.7 §4.11） ====================
  # POST /community/features 已按 M0 决议废弃（§4.7 每周自动参与 + §3 A2 服务端抽取为唯一真源）：
  # 用户一次性授权后，由服务端抽取 job 按授权状态自动纳入当周数据，客户端不再上传特征值。
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: |
        HttpOnly Session Cookie（名为 sid），登录成功后由服务端 Set-Cookie 下发，
        有效期 7 天（见 mock-server SESSION_TTL_MS）。浏览器自动携带，前端无需手动设置。
        鉴权失效时业务接口返回 401 UNAUTHENTICATED。
        （历史说明：早期契约曾假设上游 Bearer token 且不含登录/注册接口，
        实际项目改为自带 /auth/* 会话体系，本契约已据此收录，见「鉴权会话」标签。）
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: 页码，默认 1
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页条数，默认 20，上限 50
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 可选幂等键，24 小时内重复键直接返回首次结果（用于弱网下重复提交学习记录）
      schema:
        type: string
...
    RecommendationId:
      name: recommendationId
      in: path
      required: true
      description: 建议 ID
      schema:
        type: string
...
  responses:
    ValidationFailed:
      description: 参数缺失或取值越界
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: 自评专注度必须为 1-5 的整数
              field: selfReport.focus
    Unauthenticated:
      description: token 缺失或过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: 登录状态已失效，请重新登录
    GuardianAuthorizationExpired:
      description: 监护人授权失效，账号只读（PRD 8.1），仅 GET 可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: GUARDIAN_AUTHORIZATION_EXPIRED
              message: 监护人授权已过期，请重新确认后继续使用
...
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在
    RateLimited:
      description: 手动触发 AI 生成超频（PRD 6.4 控成本）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: 今日手动生成次数已达上限，请明天再试
    InternalServerError:
      description: 服务端错误。对齐 api-design-unified.md 0.2 节错误码表中的 500 INTERNAL_ERROR。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: 服务开小差了，请稍后再试
...
  schemas:
...
    Subject:
      type: string
      description: 学科代码（拼音首字母大写，与知识点库标准对齐；other 保留兜底）。YW 语文 / SX 数学 / YY 英语 / WL 物理 / HX 化学 / SW 生物 / ZZ 思想政治 / LS 历史 / DL 地理
      enum: [YW, SX, YY, WL, HX, SW, ZZ, LS, DL, other]
...
    StateLabel:
      type: string
      description: 状态标签；insufficient_data 表示数据不足不下结论
      enum: [efficient_stable, fatigue_warning, emotion_blocked, fluctuating_up, insufficient_data]
    RecScene:
      type: string
      description: 建议场景：post_session=单次学习后，weekly_review=阶段回顾，post_session_knowledge=内容维度建议（板块二 v2.2）
      enum: [post_session, weekly_review, post_session_knowledge]
    GenerationSource:
      type: string
      description: 生成来源：llm=模型生成，template=模板兜底
      enum: [llm, template]
    Rating:
      type: string
      description: 有用性评级
      enum: [useful, neutral, not_useful]

    # ---------- 鉴权会话（对齐 mock-server/server.js 与 authApi.js） ----------
...
    Error:
      type: object
      description: 统一错误格式
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 错误码
            message:
              type: string
              description: 可展示的错误说明
            field:
              type: string
              description: 校验失败的字段路径，仅参数校验错误时出现
    Pagination:
      type: object
      description: 列表响应的分页信息
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页条数
        total:
          type: integer
          description: 总条数
    GenerationStatus:
      type: object
      description: |
        AI 异步生成状态。终态为 ready / insufficient_data / failed；非终态 pending。
        两类资源的兜底语义不同（api-design-unified.md 0.3 节）：
        - 建议（Recommendation）：LLM 失败时降级为 template 兜底并仍返回 ready，
          generation.source = template。**不应出现 failed**——任何情况下用户都拿得到内容。
        - 复盘（Summary）：不做模板兜底，LLM 失败即 failed，提示稍后再试。
          **不应出现 source = template**。
        本 schema 为两者共用的超集；实现方需按资源类型约束上述互斥组合，
        调用方渲染时以资源类型 + status + source 三者共同判断。
      required: [status]
      properties:
        status:
          type: string
          description: 生成状态
          enum: [pending, ready, insufficient_data, failed]
        source:
          $ref: '#/components/schemas/GenerationSource'
        completedAt:
          type: string
          format: date-time
          description: 生成完成时间，终态时出现
...
    FeedbackRecord:
      type: object
      description: 已提交的反馈
      required: [rating, submittedAt]
      properties:
        rating:
          $ref: '#/components/schemas/Rating'
        reason:
          type: string
          nullable: true
          description: 用户补充说明
        submittedAt:
          type: string
          format: date-time
          description: 提交时间

    # ---------- 用户与设置 ----------
...
    RecommendationItem:
      type: object
      description: 单条建议
      required: [title, content]
      properties:
        title:
          type: string
          description: 建议标题
        content:
          type: string
          description: 建议正文，口语化且贴合本次学习情况
    RecommendationBasedOn:
      type: object
      description: 建议的生成依据（PRD 6.5 留痕）
      required: [stateLabel, explain]
      properties:
        assessmentId:
          type: string
          description: 依据的状态评估 ID
        recordId:
          type: string
          description: 依据的学习记录 ID
        stateLabel:
          $ref: '#/components/schemas/StateLabel'
        explain:
          type: string
          description: 面向用户的依据说明
    Recommendation:
      type: object
      description: 建议详情；生成中时 items 为 null，兜底时 generation.source 为 template 但结构一致
      required: [recommendationId, scene, generation]
      properties:
        recommendationId:
          type: string
          description: 建议 ID
        scene:
          $ref: '#/components/schemas/RecScene'
        subject:
          allOf:
            - $ref: '#/components/schemas/Subject'
          nullable: true
          description: 关联学科
        generation:
          $ref: '#/components/schemas/GenerationStatus'
        items:
          type: array
          nullable: true
          description: 建议内容；生成中为 null
          items:
            $ref: '#/components/schemas/RecommendationItem'
        basedOn:
          $ref: '#/components/schemas/RecommendationBasedOn'
        feedback:
          allOf:
            - $ref: '#/components/schemas/FeedbackRecord'
          nullable: true
          description: 用户反馈；未反馈时为 null
    RecommendationPending:
      type: object
      description: 手动请求建议后的受理响应
      required: [recommendationId, scene, generation, createdAt]
      properties:
        recommendationId:
          type: string
          description: 建议 ID，用于轮询
        scene:
          $ref: '#/components/schemas/RecScene'
        subject:
          allOf:
            - $ref: '#/components/schemas/Subject'
          nullable: true
          description: 关联学科
        generation:
          $ref: '#/components/schemas/GenerationStatus'
        createdAt:
          type: string
          format: date-time
          description: 创建时间
    RecommendationCreate:
      type: object
      description: 手动请求生成建议请求体
      required: [scene]
      properties:
        scene:
          $ref: '#/components/schemas/RecScene'
        subject:
          allOf:
            - $ref: '#/components/schemas/Subject'
          description: 可选，post_session 场景建议必填
        recordId:
          type: string
          description: 可选，post_session 场景关联的学习记录
    RecommendationList:
      type: object
      description: 建议列表响应
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    RecommendationContent:
      type: object
      required: [eligible, recordCount, recentWindowDays, reason, fromLLM]
      properties:
        eligible: { type: boolean }
        recordCount: { type: integer }
        recentWindowDays: { type: integer }
        subject:
          $ref: '#/components/schemas/Subject'
        topic: { type: string, nullable: true }
        reason: { type: string }
        fromLLM: { type: boolean }

    # ---------- 板块三：群体匿名参照（M1 契约转正） ----------
...